+++ src/Revit/dotnet/RevitExtensionDemo/Forma/Responses/GetTopFolders/Responses.cs
@@ -1,4 +1,4 @@
-namespace RevitExtensionDemo.ACC.Responses.GetTopFolders;
+namespace RevitExtensionDemo.Forma.Responses.GetTopFolders;
 
 public class Attributes
 {
@@ -110,4 +110,3 @@ public class WebView
 {
     public string Href { get; set; }
 }
-
